Possible fix for "level 6 verification error"
Upon completing level 5 in verification users get an error when attempting to verify.
According to the logs this is caused by the verification level check, in line 319-320 of lobe/views/verification/verification.py .
Checking whether the verification level the user has is indeed in the defined verification levels should stop the error from occurring. 
This may also occur for other achievements so adding it to the other checks in the future might be wise.